Week6/pyproject.toml
[tool.poetry]
name = "week6"
version = "0.1.0"
description = ""
authors = ["lucia <thisislucia1991@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
mysql-connector-python = "^8.4.0"
bcrypt = "^4.1.3"
starlette = "^0.37.2"
itsdangerous = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
